crawler/requirements.txt
# SES Job Crawler - Python Module

httpx>=0.25.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
playwright>=1.40.0
pydantic>=2.5.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
asyncio-throttle>=1.0.0
